requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
dect @ git+https://github.com/aidos-lab/dect.git
docker-pycreds==0.4.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.6.0
fsspec==2025.3.2
geotorch==0.3.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lightning==2.5.0.post0
lightning-utilities==0.14.1
littleutils==0.2.4
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.4.3
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numpy==2.1
ogb==1.3.6
outdated==0.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pdoc==15.0.3
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
rdkit==2024.9.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentry-sdk==2.28.0
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
statsmodels==0.14.5
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.6.0
torch-geometric==2.6.1
torchaudio==2.6.0
torchmetrics==1.6.3
torchvision==0.21.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.4.0
wandb==0.19.11
wcwidth==0.2.13
yarl==1.20.0
